# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ProviEmplea API",
        "description": "API REST para la plataforma de empleo ProviEmplea. Permite gestionar talentos, empresas y procesos de selección con curriculum ciego.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Servidor de desarrollo local"
        }
    ],
    "paths": {
...
        "/empresas": {
            "get": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Listar empresas activas",
                "description": "Retorna empresas activas con filtros opcionales. Límite: 60 requests/minuto.",
                "operationId": "getEmpresas",
                "parameters": [
                    {
                        "name": "tipo_empresa",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "contratacion-directa",
                                "est",
                                "outsourcing"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Empresa"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Demasiadas solicitudes. Límite: 60 req/min. Espere el tiempo indicado en el header Retry-After."
                    }
                }
            },
            "post": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Registrar nueva empresa",
                "operationId": "createEmpresa",
                "responses": {
                    "201": {
                        "description": "Empresa creada"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "/empresas/{id}": {
            "get": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Obtener empresa por ID",
                "operationId": "getEmpresa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa encontrada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Actualizar empresa",
                "operationId": "updateEmpresa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa actualizada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Desactivar empresa",
                "operationId": "deleteEmpresa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa desactivada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            }
        },
        "/empresas/{id}/validar": {
            "patch": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Validar empresa",
                "operationId": "validarEmpresa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa validada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Empresa": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "nombre_empresa": {
                        "type": "string",
                        "example": "TechCorp SpA"
                    },
                    "rut_empresa": {
                        "type": "string",
                        "example": "76123456-7"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "tipo_empresa": {
                        "type": "string",
                        "example": "contratacion-directa"
                    },
                    "contacto_nombre": {
                        "type": "string",
                        "example": "Ana López"
                    },
                    "contacto_email": {
                        "type": "string",
                        "format": "email"
                    },
                    "validado": {
                        "type": "boolean",
                        "example": false
                    },
                    "activo": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "rateLimitHeaders": {
                "type": "apiKey",
                "name": "X-RateLimit-Limit",
                "in": "header"
            }
        }
    },
...
}